rename zfits.py -> sst1meventsource.py (tests accordingly)
