+++ ctapipe/io/eventsourcefactory.py
@@ -4,7 +4,7 @@ from ctapipe.io.eventsource import EventSource
 # EventFileReader imports so that EventFileReaderFactory can see them
 # (they need to exist in the global namespace)
 import ctapipe.io.hessioeventsource
-from .import zfits
+from . import sst1meventsource
 
 __all__ = ['EventSourceFactory', 'event_source']
