+++ ctapipe/io/sst1meventsource.py
